Starter files / Start of buildNav
Here I've moved all of the starter files into this repo, designated to the Landing Page project.

I've made a start on the populating the navbar by declaring some variables and writing the buildNav function. I've just console logged hello for every section, to make sure I have the correct amount (3).

I'll continue to work on the functionality in the next commit.