Nav Bar
Here I've made the function to populate the navbar with 4 links. The function runs with the event listener domContentLoaded.

I've changed the styling in css so that the nav links are spread evenly across the width of the container. I did this by making the <ul> element (which is housing the <li> elements a flex-container.