Active class on scroll
Here I've written the functionality for the 'your-active-class' class to be added to each section when it's being viewed by the user.

Firstly I wrote the topProp function, which returns a number of how far away a section is from the top of the view-port window.

Next I wrote the addActiveClass / removeActiveClass functions. These add/remove the 'your-active-class' class from a section.

Lastly I wrote the sectionCheck function. This loops through all of the sections (defined earlier as a global variable) and gets a number for each one representing how far they are from the top of the viewport using the topProp function.

Then, using this number, a ternary operator checks if the number is between -300 and 400 (in view for the user). If it is, the addActiveClass function is used to give it the 'your-active-class' class. If it isn't the removeActiveClass is used to make sure it doesn't have the class applied.

I then added a scroll event to the window, activating the sectionCheck function when the user scrolls.